Extract dockerfile module
